% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calct}
\alias{calct}
\title{Calculate Score Test}
\usage{
calct(M1, M2)
}
\arguments{
\item{M1}{matrix with dimensions number of cases by J variants (double)}

\item{M2}{matrix with dimensions number of controls by J variants (double)}
}
\value{
calct test stat (double)
}
\description{
This function is called from the \code{test_EGPv} function in the permutation step.
}
\seealso{
\code{\link{test_EGPv}}
}

