% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{EM_calc}
\alias{EM_calc}
\title{EM calculation}
\usage{
EM_calc(M)
}
\arguments{
\item{M}{matrix of likelihoods derived from PL data.}
}
\value{
a vector of three values: p0, q0, 1-p0-q0
}
\description{
Given a n by 3 matrix, M containing likelihoods this function uses the EM algorithm to estimate genotype frequencies.
}
\details{
This part is the new part of the analysis where we can use reads instead of hard calls.\cr
For this to work we need to estimate genotype frequencies.

Data consists of matrix n by 3\cr
p = P(G=0)\cr
q = P(G=1)\cr
}

