% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{generate_seqdata_null}
\alias{generate_seqdata_alt}
\alias{generate_seqdata_null}
\title{Generate sequence data under null or alt hypothesis}
\usage{
generate_seqdata_null(L, ncase, ncont, mdcase, sddcase, mdcont, sddcont, em,
  esd, mmaf)

generate_seqdata_alt(L, ncase, ncont, mdcase, sddcase, mdcont, sddcont, em, esd,
  mmafCa, mmafCo)
}
\arguments{
\item{L}{number of variants (integer) >0}

\item{ncase}{number of cases (integer) >0}

\item{ncont}{number of controls (integer) >0}

\item{mdcase}{average read depth in cases (double) >0}

\item{sddcase}{standard deviation of read depth in cases  (double) >0}

\item{mdcont}{average read depth in controls (double) >0}

\item{sddcont}{standard deviation of read depth in controls (double) >0}

\item{em}{average error rate, probability that sequence call is wrong (double) >0}

\item{esd}{standard deviation of error rate (double) >0}

\item{mmaf}{MAF for L variants, HWE is assumed >0}

\item{mmafCa}{MAF for L variants in cases, HWE is assumed >0}

\item{mmafCo}{MAF for L variants in controls, HWE is assumed >0}
}
\value{
For null, outputs an object:

MM - conditional expected value E(Gij|Dij)\cr
P  - estimated genotype frequencies P(G=0), P(G=1), P(G=2)\cr
G  - true genotype from which sequence data generated

For alt, outputs same object with additional:

t=MM1 - conditional expected value E(Gij|Dij)

Note to Andriy: The seq_call() function that is called from  generate_seqdata_alt(), doesn't have input args: gen_freq and R.
I commented these two args out for now.  Also, gen_vect is an empty arg; see the seq_call() code.  --Ted.
}
\description{
This function is designed to mimic low and high read depth data given parameters that simulate hard calls.
}
\section{Functions}{
\itemize{
\item \code{generate_seqdata_alt}: With MAF for cases/controls.
}}

