% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{get_EG}
\alias{get_EG}
\alias{get_EG2}
\title{Conditional Expected value of genotype}
\usage{
get_EG(M, p, rdv)

get_EG2(M, p, rdv)
}
\arguments{
\item{M}{genotype likelihoods AA, Aa, aa, matrix sum(rdv) by 3 (double); uses output from get_M function.}

\item{p}{genotype frequencies AA, Aa, aa (double); output from EM_calc function.}

\item{rdv}{read depth (vector of integers) for all samples}
}
\value{
EG a vector containing conditional expectation values. Size should be equal to length(rdv).
}
\description{
This function calculates the conditional expected value (or conditional variance value) of the genotype given the genotype likelihoods and frequencies.
}
\section{Functions}{
\itemize{
\item \code{get_EG2}: but \code{get_EG2} computes the Conditional Variance value of the genotype.
}}

