% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{get_l}
\alias{get_l}
\title{Parse Phred-scaled likelihoods}
\usage{
get_l(M)
}
\arguments{
\item{M}{matrix M[i,j] of matrix, ie. a single data unit of format: GT:AD:DP:GQ:PL.  Example: get_l('0/0:2,0:4:6:0,6,42')}
}
\value{
a vector of three values: l0, l1, l2
}
\description{
This function parses the data column (GT:AD:DP:GQ:PL) to get the PL values \code{-10log10(x)},
and converts them to a vector of likelihoods: L0,L1,L2. It's called within get_Likelihood_vcf().
}
\seealso{
\code{\link{get_Likelihood_vcf}} for parent function.
}
\keyword{get_Likelihood_vcf}

