% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{get_Likelihood_vcf}
\alias{get_Likelihood_vcf}
\title{Gets Likelihoods}
\usage{
get_Likelihood_vcf(M)
}
\arguments{
\item{M}{The vcf file without headers in a matrix.}
}
\value{
Outputs an R object with:\cr
1. REF allele (vector)\cr
2. ALT allele (vector)\cr
3. POS of variant location (vector)\cr
4. FILTER indicator for 'PASS' (vector)\cr
5. likelihoods: L(D|0), L(D|1), L(D|2). List of 3 vectors.

L0 - matrix of P(D|0), rows are individuals and columns are SNPS\cr
L1 - matrix of P(D|1), rows are individuals and columns are SNPS\cr
L2 - matrix of P(D|2), rows are individuals and columns are SNPS\cr
}
\description{
This function fetches the genotype likelihoods from the VCF file (wo header) for the analysis and
must contain a 9th column (ie. 'FORMAT' field) consisting of (GT:AD:DP:GQ:PL).
}
\details{
The first step is to read the vcf file as a matrix and outputs an R object with 5 lists:

Input matrix:\cr
2nd column (POS)    - location\cr
4th column (REF)    - reference allele\cr
5th column (ALT)    - alternative allele\cr
7th column (FILTER) - vcr filter value\cr
10th column and onwards are 'data' columns in format (GT:AD:DP:GQ:PL)

PL = Phred-scaled likelihoods for genotypes as defined in the VCF specification format of 1000 Genome project.\cr
Phred-scaled likelihoods: log10(L(D|0)), log10(L(D|1)), log10(L(D|2)).

Example: 0/0:2,0:4:6:0,6,42
}

