% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{get_Lsingle}
\alias{get_Lsingle}
\title{Get the genotype likelihood for a single sequence read}
\usage{
get_Lsingle(vect_row_read, A1, A2, error)
}
\arguments{
\item{vect_row_read}{single read, ie. ('A')}

\item{A1}{reference allele, ie. (A1-'A',A2-'C')}

\item{A2}{reference allele, ie. (A1-'A',A2-'C')}

\item{error}{sequencing error (double)}
}
\value{
p - likelihood of A1 and A2 for single read
}
\description{
This function computes the genotype likelihood for a single base read given the sequencing error rate.
}
\details{
get_Lsingle is called from the \code{get_M} and \code{get_Mr} functions.
}
\seealso{
\code{\link{get_M}} or \code{\link{get_Mr}}
}

