% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{get_M}
\alias{get_M}
\alias{get_Mr}
\title{Get likelihood P(Di|G) matrix}
\usage{
get_M(Error, vect_row_reads)

get_Mr(Error, vect_row_reads, rdv)
}
\arguments{
\item{Error}{vector of base call error rates; Uses the same \code{error} input in the \code{seq_call} function but concatenate all \code{error} values over all samples.}

\item{vect_row_reads}{vector of row sequence reads ('A','G','T','C') outputted from function \code{seq_call}, but concatenated over all samples in savme fashion as \code{Error}.}

\item{rdv}{vector of read depths concatenated for all samples (integer).   Looks like getting likelihood over all reads cumulatively? Andriy, please explain.}
}
\value{
M matrix of P(read|AA), P(read|Aa) and P(read|aa), where row dimensions equal \code{length(Error)}, by 3 columns.
}
\description{
This function calculates a likelihood matrix given a vector of error rates and a vector of sequence reads.
}
\details{
Called inside the \code{generate_seqdata} functions.  The output of this function \code{get_M} are used as inputs to \code{get_EG}.
Note to Andriy:  Why is LG hardcoded to LG=c('TT','CT','CC')  ??  Guess it's to denote the AA, Aa, aa.
}
\section{Functions}{
\itemize{
\item \code{get_Mr}: with \code{rdv}
}}
\seealso{
\code{\link{seq_call}}
}

