% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{seq_call}
\alias{get_row_reads}
\alias{seq_call}
\title{Produces base calls}
\usage{
seq_call(genotype, error, gen_vect)

get_row_reads(error, genotype, depth)
}
\arguments{
\item{genotype}{vector of two alleles for a given genotype, ie. c('C', 'T')}

\item{error}{vector of base call errors: Phen scores (double).  Length of this vector is the read depth.}

\item{gen_vect}{vector of prior genotypes.}

\item{depth}{the read depth (integer)}
}
\value{
vect_row_reads vector of row sequence reads ('A','G','T','C') for single variant. The length of this vector is the read depth.
}
\description{
This function produces the base call from given a genotype and the error scores. It is called from the \code{generate_seqdata} functions.
}
\section{Functions}{
\itemize{
\item \code{get_row_reads}: see \code{seq_call} is the wrapper around get_row_reads.
}}
\seealso{
\code{\link{get_row_reads}}

\code{\link{seq_call}}
}

