% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{test_EG}
\alias{test_EG}
\alias{test_EGPv}
\title{Regular Score/Trend test}
\usage{
test_EG(M1, M2)

test_EGPv(M1, M2, P, perm)
}
\arguments{
\item{M1}{matrix with dimensions number of cases by J variants (double)}

\item{M2}{matrix with dimensions number of controls by J variants (double)}

\item{P}{a J by 3 matrix (double) of estimated genotype frequency for J variants: P(G=0), P(G=1) and P(G=2).}

\item{perm}{number of permutations (int)}
}
\value{
vector of p-values for J variants (double)
}
\description{
The test_EG function is a regular score/trend test evaluated by asymptotic distribution.

The test_EGPv function is a regular score/trend test evaluated by permutation distribution.
}
\details{
The inputs are two matrices of genotypes for cases and controls.  The matrix should have dimensions:
number of samples (\code{ncase} or \code{ncont}) by number of variants \code{J}.

\code{test_EG} - Regular Score/Trend test (asymptotic distbn).\cr
\code{test_EGPv} - Regular Score/Trend test (permutation distbn).
}
\section{Functions}{
\itemize{
\item \code{test_EGPv}: 
}}

