% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{test_EGC}
\alias{test_EGB}
\alias{test_EGBnew}
\alias{test_EGC}
\alias{test_EGCV}
\title{Robust Variance Score (RVS)}
\usage{
test_EGC(M1, M2, P)

test_EGCV(M1, M2, P)

test_EGB(M1, M2, P, perm)

test_EGBnew(M1, M2, P, perm)
}
\arguments{
\item{M1}{matrix with dimensions of number of cases by J variants (double)}

\item{M2}{matrix with dimensions of number of controls by J variants (double)}

\item{P}{matrix with dimensions J by 3 that consists of estimated genotype frequency for J variants: P(G=0), P(G=1) and P(G=2). (double)}

\item{perm}{number of permutations (int)}
}
\value{
vector of p-values for J variants (double)
}
\description{
The test_EGC function is the RVS test evaluated by asymptotic distribution.

The test_EGCV function is the RVS test evaluated by asymptotic distribution using the Regular Variance Estimates for cases and controls.

The test_EGB function is the RVS test evaluated by bootstrap permutation using the Robust Variance Estimates for cases and controls.

The test_EGBnew function is the RVS test evaluated by bootstrap distribution using the Regular Variance Estimates for cases and controls.
}
\details{
RVS uses the Robust Variance Estimates described in the journal Bioinformatics:
Vol.30 no.15 2014, pp.2179 to 2188.

The input values are matrices of expected values of genotypes given sequence data for cases and controls.

\code{test_EGC} - RVS (asymptotic distbn).\cr
\code{test_EGCV} - RVS using Regular Variance Estimates for cases/ctrls (asymptotic distn).\cr
\code{test_EGB} - RVS using Robust Variance Estimates for cases/ctrls (bootstrap perm).\cr
\code{test_EGBnew} - RVS using the Regular Variance Estimates for cases/ctrls (bootstrap distn and perm).
}
\section{Functions}{
\itemize{
\item \code{test_EGCV}: 

\item \code{test_EGB}: 

\item \code{test_EGBnew}: 
}}

